Add supplemental analysis and figs to chapter 2
Added supplemental analysis that I originally performed as part of the
Lee and Huddleston et al. 2018 project and that led to us using the
mutational preferences from individual clades instead of full-length
sequence preferences. Also, added the logo plot of DMS preferences for
H3N2 and the corresponding table with the rescaling parameter for
completeness.